requirements.txt
torch~=1.13.1
Pillow~=9.5.0
torchvision~=0.14.1
numpy~=1.21.6
scipy~=1.10.1
matplotlib~=3.7.1
opencv-python~=4.1.2.30
boto3~=1.26.163
Flask~=2.3.2
requests~=2.31.0
botocore~=1.29.163
scikit-image~=0.21.0
insightface~=0.2.1